bug fixing,refactoring, and add objects for OOP_DZ2

+++ src/OOP_DZ/DZ2/Clases/WildAnimal.java
@@ -9,14 +9,15 @@ public abstract class WildAnimal extends Animal implements Hunting {
     String habitat; //Место обитания
     LocalDate locationDate;
 
-    public WildAnimal(int animalHeight, int animalWeight, String animalEyeColor, String habitat, LocalDate locationDate) {
+    public WildAnimal(double animalHeight, double animalWeight, String animalEyeColor, String habitat, LocalDate locationDate) {
         super(animalHeight, animalWeight, animalEyeColor);
         this.habitat = habitat;
         this.locationDate = locationDate;
     }
 
+
     @Override
     public String info() {
-        return super.info()+", I live in "+habitat +" since " + locationDate;
+        return super.info() + ", I live in " + habitat + " since " + locationDate;
     }
 }
